refactor(bridge): extract reaction events
